Reverted r9614 and r9615.
Sorry, bct, but I think it is to early to merge that as it completely
breaks. It seems nobody in gajim@conference.gajim.org considers it
usable yet.
I don't know if you got asterix' ok for it and I'm sorry if I reverted
it now although you had his ok, but having broken trunk is very
contra-productive. I think it was just too early to merge.


git-svn-id: svn://svn.gajim.org/gajim/trunk@9616 c4d4e39b-2f5f-9a77-1c94-36969bb79148