Partially fix #4098.
This is, however, only a half fix as this fixes that it fails when
there's None in waiting_descriptors[2], but actually, there never
should be None, so I have to investigate why there was none.

This patch is, however, correct and needed and the other queues also
have that check. It's just that I also need to fix the reason for
the None in the queue.


git-svn-id: svn://svn.gajim.org/gajim/trunk@9941 c4d4e39b-2f5f-9a77-1c94-36969bb79148